dashboard.html modded to dashboard.php
Changed dashboard.html to dashboard.php to support user form submmission. This is temporary for debugging so can change the storeid on the fly so that BA side need not touch the code to play around.
wordcloud.js now takes in the parameter passed by the user to display the wordcloud for selected store.