WordCloud - 2Bug Fixes and 1 Improvement
Bug Fix 1:
Word cloud can handle 1,2 or 3 adjs now.

Bug Fix 2:
Store 9 had just 1 adj and log(1) is 0 resulting in an infinite while loop. now if total adj value is 1 returns 2 so log(2) will be the smallest size.

Improvement:
Font size are based on the average count of all all the adjs in total. This is so that shops with smaller number of adj dont have a substantially smaller font size and they can all be roughly the same.