Overall- Added Spinners
Spinners will allow the user to be aware that the wordcloud/sentiment charts are loading.
WordCloud has its own individual spinner as user will be clicking on the sentiment donut to see wordcloud change and we dont want the whole page to be on overlay.
For the rest since they are all taking the infro from a single endpoint when the page loads an overlay will be made to cover the entire screen to show it is loading.