Sentiment Donut - Modified arcs on_click behaviour
Previously clicking a single arc will just expand it and clicking on it again shrinks it. However that use case was so that we can for example combine negative and neutral reviews to show it on the wordcloud. This would also mean if the user want to just see either positive negative or neutral reviews alone, they would need to deselect somthing that is already selected to select another arc.

Thus now when one arc is selected e.g. Positive and when another arc is selected e.g. Negative then the Positive will be automatially deselected.