URADashboard - Stores Ranked
A new approach to how the stores rank is displayed.

Previously upon clicking each store in the rank, it will change
1. Overall Sentiment Score Stars
2. Total Reviews
3. Sentiment Over Time Stacked Bar Chart
4. Word Cloud
by updating these information to the store that was clicked.

The new approach is upon clicking each store, the Overall sentiment score stars as well as the total number of reivew this store has appears just below the name of the store clicked. And only the
1. Sentiment Over Time Stacked Bar Chart
2. Word cloud
Will update the info for the store that was clicked.