BIZDashboard - SOT Select Feature
Previously when user selects any rect for the SOT the reviews get populated but the rect selected doesnt stay selected(no strokes).

To standardize with the donut chart,where on select the selection shld have a stroke, now SOT also mimics this behaviour.
Additionally, when the user clicks on the close button for the reviews container, all strokes will be cleared in the SOT.