URADashboard - SOT Select Feature
Same feature as commit-f70170cb8127a16d25a9facbdd2726243ddc9c26 but for URADashboard

Previously when user selects any rect for the SOT the reviews get populated but the rect selected doesnt stay selected(no strokes).

To standardize with the donut chart,where on select the selection shld have a stroke, now SOT also mimics this behaviour.
Additionally, when the user clicks on the close button for the reviews container, all strokes will be cleared in the SOT.