URADashboard - Reply To Customer
URA can now directly reply to the customer from the dashboard by clicking on the reply icon.
The TO: SUBJECT: and BODY: fields will be prepopulated and open the default mail app. URA can then edit further or just send the email.
The Subject and body is now in common.js